backend/requirements.txt
asgiref==3.8.1
babel==2.17.0
contourpy==1.3.2
cycler==0.12.1
Django==4.2.23
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
docxcompose==1.4.0
docxtpl==0.20.0
fonttools==4.58.4
gunicorn==23.0.0
Jinja2==3.1.6
kiwisolver==1.4.8
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pillow==11.2.1
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
whitenoise==6.9.0
